backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
paddlepaddle-gpu==2.6.1
paddleocr>=2.8.0
opencv-python>=4.9.0
imagehash>=4.3.1
Pillow>=10.3.0
pysrt>=1.1.2
ffmpeg-python>=0.2.0
ollama>=0.2.0
httpx>=0.27.0
